started moving a bunch of packages to allow extracting the whole-body control core
significantly overhauled and set up custom pid parameters for the whole body controller core

moved a lot of the filtered yo variables into a common project that can be moved to commons

moved a bunch of the polynomial classes out after getting things solidified, but need to fix the linear systems tests

cleaned up a lot more of the yo variable types and moved their packages

got more code moved and most of the linear algebra tests fixed

removed unused code

set up wbcc for extraction

cleaned up the ihmc math package for extraction

prepped the yo variables by moving them into their own packages

did considerable package movement, but solidified how the extracted repos will be structured

moved out more of th e simple code into the sub-sets

started refactoring some of the code to prep for the commons move

set up the euclid update repo, including javadoc and pulling over the tests

moved the mecano packages around

set up a robotics tools project to become its own repo

set up the ihmc math library more properly, and worked to add some javadoc on some of the core classes

did a bunch of testing and javadoc, and cleaned up polynomial3D

moved all the code into the proper packages

fixed some formatting and a broken test
