--- ihmc-robotics-toolkit/src/main/java/us/ihmc/robotics/time/TimeIntervalProvider.java
@@ -1,6 +1,0 @@
-package us.ihmc.robotics.time;
-
-public interface TimeIntervalProvider
-{
-   TimeIntervalBasics getTimeInterval();
-}
