+++ src/main/java/us/ihmc/commons/time/TimeIntervalBasics.java
@@ -22,7 +22,7 @@ public interface TimeIntervalBasics extends TimeIntervalReadOnly
     */
    default void reset()
    {
-      setInterval(Double.NaN, Double.NaN);
+      setIntervalUnsafe(Double.NaN, Double.NaN);
    }
 
    /**
@@ -30,9 +30,14 @@ public interface TimeIntervalBasics extends TimeIntervalReadOnly
     */
    default void setInterval(double startTime, double endTime)
    {
+      setIntervalUnsafe(startTime, endTime);
+      checkInterval();
+   }
+
+   default void setIntervalUnsafe(double startTime, double endTime)
+   {
       setStartTime(startTime);
       setEndTime(endTime);
-      checkInterval();
    }
 
    /**
@@ -49,7 +54,7 @@ public interface TimeIntervalBasics extends TimeIntervalReadOnly
     */
    default TimeIntervalBasics shiftInterval(double shiftTime)
    {
-      setInterval(getStartTime() + shiftTime, getEndTime() + shiftTime);
+      setIntervalUnsafe(getStartTime() + shiftTime, getEndTime() + shiftTime);
       return this;
    }
